time-agent/pyproject.toml
[project]
name = "time-agent"
version = "1.0.0"
description = "Time and Timezone Agent with A2A Protocol v0.4.0 Architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk==0.2.8",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain-openai>=0.1.0",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "starlette>=0.27.0",
    "mcp>=1.0.0",
    "pytz>=2024.1",
    "tzdata>=2024.1",
]

[project.scripts]
time-agent = "src.time_agent.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true